requirements_p.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.4
pyyaml==6.0.1

# Deep learning & detection
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.220
easyocr==1.7.1

# Toxicity detection (compatible versions)
detoxify==0.5.1
transformers==4.22.1  # <- downgraded to match Detoxify’s requirement

# Web interface
streamlit==1.28.2
gradio==4.7.1

# Utilities
pillow==10.1.0
tqdm==4.66.1

# Development & testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
